backend/requirements.txt
flask
flask-cors
yfinance
pandas
requests
pytest
python-dotenv
schwab-py
